Attempt on Resolving Merge Issues
